backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
psycopg2-binary==2.9.9
pandas==2.1.4
django-cors-headers>=4.3,<5.0
pytest==7.4.3
pytest-django>=4.7,<5.0
gunicorn>=21.2
bcrypt>=4.0.0
PyYAML>=6.0
django-environ>=0.11.2
structlog>=24.1.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5.0,<3.0
whitenoise>=6.6.0,<7.0
django-prometheus>=2.3.1,<3.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-django>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
